Renamed some hooks, created basic new project form and fixed useHistory bugs
Some hooks did not have the best names to explain what is their exact job.
New project form created and upon load of homepage user is displayed with New Project form.
User was able to click a link multiple times and push that url to history. To avoid this I added a if statement to check if user is trying to push the current link to history