Fully implemented New Project process & removed console.log()'s
Had to edit getCurrentProjectUsers() in Home.js to make sure that it would return an empty array if filtering found nothing.
This was creating an error after a user made a new project and was redirected to the new project page. The error was due to the allProjects array not containing the latest projects.